+++ tests/test_create_schema.py
@@ -2,8 +2,9 @@
 
 import unittest
 import os
-import jello.cli
-from jello.cli import opts
+import jello
+from jello.options import opts
+from jello.lib import set_env_colors, create_schema
 
 
 class MyTests(unittest.TestCase):
@@ -26,13 +27,13 @@ class MyTests(unittest.TestCase):
         opts.arraybracket_color = None
 
         # initialize schema_list
-        jello.cli.schema_list = []
+        jello.lib.schema_list = []
 
         # initialize JELLO_COLORS env variable
         os.environ['JELLO_COLORS'] = 'default,default,default,default,default,default'
 
         # set the colors
-        jello.cli.set_env_colors()
+        set_env_colors()
 
         # create samples
         self.dict_sample = {
@@ -127,8 +128,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = True
         self.expected = '. = \x1b[90mtrue\x1b[0m;'
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     def test_true_m(self):
         """
@@ -137,8 +138,8 @@ class MyTests(unittest.TestCase):
         self.data_in = True
         self.expected = '. = true;'
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     #
     # Naked False
@@ -150,8 +151,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = False
         self.expected = '. = \x1b[90mfalse\x1b[0m;'
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     def test_false_m(self):
         """
@@ -160,8 +161,8 @@ class MyTests(unittest.TestCase):
         self.data_in = False
         self.expected = '. = false;'
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     #
     # Naked null
@@ -173,8 +174,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = None
         self.expected = '. = \x1b[90mnull\x1b[0m;'
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     def test_null_m(self):
         """
@@ -183,8 +184,8 @@ class MyTests(unittest.TestCase):
         self.data_in = None
         self.expected = '. = null;'
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     #
     # naked int
@@ -196,8 +197,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = 42
         self.expected = '. = \x1b[35m42\x1b[0m;'
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     def test_int_m(self):
         """
@@ -206,8 +207,8 @@ class MyTests(unittest.TestCase):
         self.data_in = 42
         self.expected = '. = 42;'
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     #
     # naked float
@@ -219,8 +220,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = 3.14
         self.expected = '. = \x1b[35m3.14\x1b[0m;'
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     def test_float_m(self):
         """
@@ -229,8 +230,8 @@ class MyTests(unittest.TestCase):
         self.data_in = 3.14
         self.expected = '. = 3.14;'
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     #
     # naked string
@@ -242,8 +243,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = '"string with\\nnewline char"'
         self.expected = '. = \x1b[32m"\\"string with\\\\nnewline char\\""\x1b[0m;'
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     def test_string_m(self):
         """
@@ -252,8 +253,8 @@ class MyTests(unittest.TestCase):
         self.data_in = '"string with\\nnewline char"'
         self.expected = '. = "\\"string with\\\\nnewline char\\"";'
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual('\n'.join(jello.cli.schema_list), self.expected)
+        create_schema(self.data_in)
+        self.assertEqual('\n'.join(jello.lib.schema_list), self.expected)
 
     #
     # Naked Dict
@@ -265,8 +266,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = self.dict_sample
         self.expected = ['.\x1b[1m\x1b[34mstring\x1b[0m = \x1b[32m"string\\nwith newline\\ncharacters in it"\x1b[0m;', '.\x1b[1m\x1b[34mtrue\x1b[0m = \x1b[90mtrue\x1b[0m;', '.\x1b[1m\x1b[34mfalse\x1b[0m = \x1b[90mfalse\x1b[0m;', '.\x1b[1m\x1b[34mnull\x1b[0m = \x1b[90mnull\x1b[0m;', '.\x1b[1m\x1b[34mint\x1b[0m = \x1b[35m42\x1b[0m;', '.\x1b[1m\x1b[34mfloat\x1b[0m = \x1b[35m3.14\x1b[0m;', '.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[32m"string\\nwith newline\\ncharacters in it"\x1b[0m;', '.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mtrue\x1b[0m;', '.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m2\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mfalse\x1b[0m;', '.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m3\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mnull\x1b[0m;', '.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m4\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m42\x1b[0m;', '.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m5\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m3.14\x1b[0m;']
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_dict_m(self):
         """
@@ -275,8 +276,8 @@ class MyTests(unittest.TestCase):
         self.data_in = self.dict_sample
         self.expected = ['.string = "string\\nwith newline\\ncharacters in it";', '.true = true;', '.false = false;', '.null = null;', '.int = 42;', '.float = 3.14;', '.array[0] = "string\\nwith newline\\ncharacters in it";', '.array[1] = true;', '.array[2] = false;', '.array[3] = null;', '.array[4] = 42;', '.array[5] = 3.14;']
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     #
     # true in a list
@@ -288,8 +289,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = [True]
         self.expected = ['.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mtrue\x1b[0m;']
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_list_true_m(self):
         """
@@ -298,8 +299,8 @@ class MyTests(unittest.TestCase):
         self.data_in = [True]
         self.expected = ['.[0] = true;']
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     #
     # false in a list
@@ -311,8 +312,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = [False]
         self.expected = ['.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mfalse\x1b[0m;']
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_list_false_m(self):
         """
@@ -321,8 +322,8 @@ class MyTests(unittest.TestCase):
         self.data_in = [False]
         self.expected = ['.[0] = false;']
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     #
     # null in a list
@@ -334,8 +335,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = [None]
         self.expected = ['.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mnull\x1b[0m;']
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_list_null_m(self):
         """
@@ -344,8 +345,8 @@ class MyTests(unittest.TestCase):
         self.data_in = [None]
         self.expected = ['.[0] = null;']
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     #
     # Int in a list
@@ -357,8 +358,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = [42]
         self.expected = ['.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m42\x1b[0m;']
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_list_int_m(self):
         """
@@ -367,8 +368,8 @@ class MyTests(unittest.TestCase):
         self.data_in = [42]
         self.expected = ['.[0] = 42;']
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     #
     # Float in a list
@@ -380,8 +381,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = [3.14]
         self.expected = ['.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m3.14\x1b[0m;']
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_list_float_m(self):
         """
@@ -390,8 +391,8 @@ class MyTests(unittest.TestCase):
         self.data_in = [3.14]
         self.expected = ['.[0] = 3.14;']
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     #
     # String in a list
@@ -403,8 +404,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = ['string with spaces\nand newline\ncharacters']
         self.expected = ['.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[32m"string with spaces\\nand newline\\ncharacters"\x1b[0m;']
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_list_str_m(self):
         """
@@ -413,8 +414,8 @@ class MyTests(unittest.TestCase):
         self.data_in = ['string with spaces\nand newline\ncharacters']
         self.expected = ['.[0] = "string with spaces\\nand newline\\ncharacters";']
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     #
     # List with different types of elements
@@ -426,8 +427,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = self.list_sample
         self.expected = ['.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[32m"string\\nwith newline\\ncharacters in it"\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mtrue\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m2\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mfalse\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m3\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mnull\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m4\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m42\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m5\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m3.14\x1b[0m;']
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_list_sample_m(self):
         """
@@ -436,8 +437,8 @@ class MyTests(unittest.TestCase):
         self.data_in = self.list_sample
         self.expected = ['.[0] = "string\\nwith newline\\ncharacters in it";', '.[1] = true;', '.[2] = false;', '.[3] = null;', '.[4] = 42;', '.[5] = 3.14;']
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     #
     # Dicts in a list
@@ -449,8 +450,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = self.list_of_dicts_sample
         self.expected = ['.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mstring\x1b[0m = \x1b[32m"string\\nwith newline\\ncharacters in it"\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mtrue\x1b[0m = \x1b[90mtrue\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mfalse\x1b[0m = \x1b[90mfalse\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mnull\x1b[0m = \x1b[90mnull\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mint\x1b[0m = \x1b[35m42\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mfloat\x1b[0m = \x1b[35m3.14\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[32m"string\\nwith newline\\ncharacters in it"\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mtrue\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m2\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mfalse\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m3\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mnull\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m4\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m42\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m5\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m3.14\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mstring\x1b[0m = \x1b[32m"another string\\nwith newline\\ncharacters in it"\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mtrue\x1b[0m = \x1b[90mtrue\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mfalse\x1b[0m = \x1b[90mfalse\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mnull\x1b[0m = \x1b[90mnull\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mint\x1b[0m = \x1b[35m10001\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34mfloat\x1b[0m = \x1b[32m-400.45\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[32m"string\\nwith newline\\ncharacters in it"\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mtrue\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m2\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mfalse\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m3\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mnull\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m4\x1b[0m\x1b[35m]\x1b[0m = \x1b[32m-6000034\x1b[0m;', '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34marray\x1b[0m\x1b[35m[\x1b[0m\x1b[31m5\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m999999.854321\x1b[0m;']
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_list_dict_m(self):
         """
@@ -459,8 +460,8 @@ class MyTests(unittest.TestCase):
         self.data_in = self.list_of_dicts_sample
         self.expected = ['.[0].string = "string\\nwith newline\\ncharacters in it";', '.[0].true = true;', '.[0].false = false;', '.[0].null = null;', '.[0].int = 42;', '.[0].float = 3.14;', '.[0].array[0] = "string\\nwith newline\\ncharacters in it";', '.[0].array[1] = true;', '.[0].array[2] = false;', '.[0].array[3] = null;', '.[0].array[4] = 42;', '.[0].array[5] = 3.14;', '.[1].string = "another string\\nwith newline\\ncharacters in it";', '.[1].true = true;', '.[1].false = false;', '.[1].null = null;', '.[1].int = 10001;', '.[1].float = -400.45;', '.[1].array[0] = "string\\nwith newline\\ncharacters in it";', '.[1].array[1] = true;', '.[1].array[2] = false;', '.[1].array[3] = null;', '.[1].array[4] = -6000034;', '.[1].array[5] = 999999.854321;']
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     #
     # lists in list
@@ -472,8 +473,8 @@ class MyTests(unittest.TestCase):
         """
         self.data_in = self.list_of_lists_sample
         self.expected = ['.\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m[\'string\\nwith newline\\ncharacters in it\', True, False, None, 42, 3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[32m"string\\nwith newline\\ncharacters in it"\x1b[0m;', ".\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mtrue\x1b[0m;", ".\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m2\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mfalse\x1b[0m;", ".\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m3\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mnull\x1b[0m;", ".\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m4\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m42\x1b[0m;", ".\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m5\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m3.14\x1b[0m;", '.\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m[\'another string\\nwith newline\\ncharacters in it\', True, False, None, 42001, -3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m0\x1b[0m\x1b[35m]\x1b[0m = \x1b[32m"another string\\nwith newline\\ncharacters in it"\x1b[0m;', ".\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mtrue\x1b[0m;", ".\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m2\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mfalse\x1b[0m;", ".\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m3\x1b[0m\x1b[35m]\x1b[0m = \x1b[90mnull\x1b[0m;", ".\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m4\x1b[0m\x1b[35m]\x1b[0m = \x1b[35m42001\x1b[0m;", ".\x1b[35m[\x1b[0m\x1b[31m1\x1b[0m\x1b[35m]\x1b[0m.\x1b[1m\x1b[34m['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14]\x1b[0m\x1b[35m[\x1b[0m\x1b[31m5\x1b[0m\x1b[35m]\x1b[0m = \x1b[32m-3.14\x1b[0m;"]
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
     def test_list_list_m(self):
         """
@@ -482,8 +483,8 @@ class MyTests(unittest.TestCase):
         self.data_in = self.list_of_lists_sample
         self.expected = ['.[0].[\'string\\nwith newline\\ncharacters in it\', True, False, None, 42, 3.14][0] = "string\\nwith newline\\ncharacters in it";', ".[0].['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14][1] = true;", ".[0].['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14][2] = false;", ".[0].['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14][3] = null;", ".[0].['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14][4] = 42;", ".[0].['string\\nwith newline\\ncharacters in it', True, False, None, 42, 3.14][5] = 3.14;", '.[1].[\'another string\\nwith newline\\ncharacters in it\', True, False, None, 42001, -3.14][0] = "another string\\nwith newline\\ncharacters in it";', ".[1].['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14][1] = true;", ".[1].['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14][2] = false;", ".[1].['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14][3] = null;", ".[1].['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14][4] = 42001;", ".[1].['another string\\nwith newline\\ncharacters in it', True, False, None, 42001, -3.14][5] = -3.14;"]
         opts.mono = True
-        jello.cli.create_schema(self.data_in)
-        self.assertEqual(jello.cli.schema_list, self.expected)
+        create_schema(self.data_in)
+        self.assertEqual(jello.lib.schema_list, self.expected)
 
 
 if __name__ == '__main__':
